+++ tests/testthat/test_relatedness.R
@@ -26,16 +26,35 @@ test_that("asTasselDistanceMatrix() returns correct data and exceptions", {
 test_that("mds() returns correct data and exceptions", {
     expect_error(mds(tasGeno))
 
-    # set.seed(123)
-    # m <- 10
-    # s <- matrix(rnorm(100), m)
-    # s[lower.tri(s)] <- t(s)[lower.tri(s)]
-    # diag(s) <- 2
-    # colnames(s) <- rownames(s) <- paste0("s_", seq_len(m))
-    #
-    # sT <- distanceMatrix(tasGeno)
-    #
-    # expect_true(inherits(mds(sT, "list")))
+    tasDist <- distanceMatrix(tasGeno)
+    mdsRes  <- mds(tasDist)
+
+    expect_s4_class(mdsRes, "MDSResults")
+    expect_equal(
+        reportNames(mdsRes),
+        c("MDS_PCs_Datum", "MDS_Eigenvalues_Datum")
+    )
+    expect_equal(
+        colnames(tableReport(mdsRes)),
+        c("Taxa", paste0("PC", 1:5))
+    )
+    expect_equal(
+        nrow(tableReport(mdsRes)),
+        length(taxaList(tasGeno))
+    )
+    expect_equal(
+        nrow(tableReport(mdsRes, "MDS_Eigenvalues_Datum")),
+        5
+    )
+    expect_equal(
+        rJava::.jclass(mdsRes@jObj),
+        "net.maizegenetics.phenotype.CorePhenotype"
+    )
+
+    expect_equal(
+        colnames(tableReport(mds(tasDist, nAxes = 3))),
+        c("Taxa", paste0("PC", 1:3))
+    )
 })